pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wgmesh"
version = "0.1.0"
description = "WireGuard hub-and-spoke mesh network management CLI"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.scripts]
wgmesh = "wgmesh.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/wgmesh"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
]
